--- client/ScrollButton.css
@@ -1,25 +1,0 @@
-.scroll-to-top {
-  position: fixed;
-  bottom: 30px;
-  right: 30px;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 50%;
-  width: 50px;
-  height: 50px;
-  font-size: 24px;
-  cursor: pointer;
-  opacity: 0;
-  transition: opacity 0.3s ease, transform 0.3s ease; 
-  z-index: 1000;
-}
-
-.scroll-to-top.visible {
-  opacity: 1;
-}
-
-.scroll-to-top:hover {
-  background-color: #0056b3;
-  transform: scale(1.2); 
-}
