+++ client/src/App.jsx
@@ -1,50 +1,51 @@
 import { Route, Routes } from "react-router-dom";
 import { AnimatePresence } from "framer-motion";
-import Home from "./Pages/Home";
-import About from "./Pages/About";
-import Models from "./Pages/Models";
-import Testimonials from "./Pages/Testimonials";
-import Navbar from "./components/default/Navbar";
-import Footer from "./components/default/Footer";
-import Team from "./Pages/Team";
-import Contact from "./Pages/Contact";
-import Errorpage from "./Pages/Errorpage";
+import Home from "./pages/Home";
+import About from "./pages/About";
+import Models from "./pages/Models";
+import Testimonials from "./pages/Testimonials";
+import Team from "./pages/Team";
+import Contact from "./pages/Contact";
+import Errorpage from "./pages/Errorpage";
 import Login from "./components/Auth/Login";
 import Register from "./components/Auth/Register";
-import LearnMore from "./Pages/LearnMore";
+import LearnMore from "./pages/LearnMore";
 import Booking from "./components/Booking/Booking";
 import ProtectedRoute from "./components/ProtectedRoute/ProtectedRoute";
+import { MainLayout } from "./layout/MainLayout";
+import { AuthLayout } from "./layout/AuthLayout";
 
 function App() {
   return (
-    <div className="min-h-screen w-full flex flex-col bg-white">
-      <Navbar />
-      <main className="flex-grow">
-        <AnimatePresence mode="wait">
-          <Routes>
-            <Route index path="/" element={<Home />} />
-            <Route path="/about" element={<About />} />
-            <Route path="/models" element={<Models />} />
-            <Route path="/testimonials" element={<Testimonials />} />
-            <Route path="/team" element={<Team />} />
-            <Route path="/contact" element={<Contact />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/register" element={<Register />} />
-            <Route path="/learnmore" element={<LearnMore />} />
-            <Route path="*" element={<Errorpage />} />
-            <Route
-              path="/booking"
-              element={
-                <ProtectedRoute>
-                  <Booking />
-                </ProtectedRoute>
-              }
-            />
-          </Routes>
-        </AnimatePresence>
-      </main>
-      <Footer />
-    </div>
+    <AnimatePresence mode="wait">
+      <Routes>
+        {/* Auth routes without Navbar and Footer */}
+        <Route element={<AuthLayout />}>
+          <Route path="/login" element={<Login />} />
+          <Route path="/register" element={<Register />} />
+        </Route>
+
+        {/* Main routes with Navbar and Footer */}
+        <Route element={<MainLayout />}>
+          <Route index path="/" element={<Home />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/models" element={<Models />} />
+          <Route path="/testimonials" element={<Testimonials />} />
+          <Route path="/team" element={<Team />} />
+          <Route path="/contact" element={<Contact />} />
+          <Route path="/learnmore" element={<LearnMore />} />
+          <Route path="*" element={<Errorpage />} />
+          <Route
+            path="/booking"
+            element={
+              <ProtectedRoute>
+                <Booking />
+              </ProtectedRoute>
+            }
+          />
+        </Route>
+      </Routes>
+    </AnimatePresence>
   );
 }
