+++ client/src/layout/AuthLayout.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+const AuthLayout = () => {
+  return (
+    <div className="min-h-screen w-full flex flex-col bg-white">
+      <main className="flex-grow">
+        <Outlet />
+      </main>
+    </div>
+  );
+};
+export { AuthLayout };
