+++ client/src/layout/MainLayout.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import Navbar from "../components/default/Navbar";
+import Footer from "../components/default/Footer";
+import { Outlet } from "react-router-dom";
+
+const MainLayout = () => {
+  return (
+    <div className="min-h-screen w-full flex flex-col bg-white">
+      <Navbar />
+      <main className="flex-grow">
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  );
+};
+
+export { MainLayout };
